requirements.txt
python-telegram-bot[job-queue]
ccxt
pandas
aiohttp
python-dotenv
requests